app.tsx comments implemented, home file added
